Move Domain classes to domain namespace
