+++ src/Web/Controller/DashboardController.php
@@ -7,7 +7,7 @@ namespace VDOLog\Web\Controller;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
-use VDOLog\Web\Entity\Game;
+use VDOLog\Core\Domain\Game;
 
 class DashboardController extends AbstractController
 {
