+++ src/Web/Controller/GameController.php
@@ -11,7 +11,7 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\StreamedResponse;
 use Symfony\Component\Routing\Annotation\Route;
-use VDOLog\Web\Entity\Game;
+use VDOLog\Core\Domain\Game;
 use VDOLog\Web\Form\GameType;
 use VDOLog\Web\Model\GameExporter;
