+++ src/Web/Form/GameType.php
@@ -9,7 +9,7 @@ use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 use Symfony\Component\Validator\Constraints\NotBlank;
-use VDOLog\Web\Entity\Game;
+use VDOLog\Core\Domain\Game;
 
 class GameType extends AbstractType
 {
